refactor(materializer): extract scene-entry-array Proxy machinery (file #3)
Lift the lazy scene-entry array views (patch/append/truncate/splice builders +
createSceneEntryArrayView, addSceneEntryPatch, toArrayIndex, createArrayLikeOwnKeys)
into scene-entry-array.ts. Pure structural helpers. Net 74/74, check 0.
